Milight: Move simulated V6 bridge to own bundle. Use OSGi DS annotations
