+++ addons/binding/org.openhab.binding.milight/OSGI-INF/.gitignore
@@ -1,0 +1,1 @@
+*.xml
