--- addons/binding/org.openhab.binding.milight/OSGI-INF/MilightDiscovery.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2010-2017 by the respective copyright holders.
-
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.binding.milight.internal.discovery.MilightBridgeDiscovery">
- <implementation class="org.openhab.binding.milight.internal.discovery.MilightBridgeDiscovery"/>
- <service>
-    <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
-  
- </service>
-</scr:component>
