+++ addons/binding/org.openhab.binding.milight/src/main/java/org/openhab/binding/milight/internal/MilightHandlerFactory.java
@@ -15,10 +15,12 @@ import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
+import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.openhab.binding.milight.MilightBindingConstants;
 import org.openhab.binding.milight.handler.MilightBridgeV3Handler;
 import org.openhab.binding.milight.handler.MilightBridgeV6Handler;
 import org.openhab.binding.milight.handler.MilightLedHandler;
+import org.osgi.service.component.annotations.Component;
 
 import com.google.common.collect.Sets;
 
@@ -28,6 +30,7 @@ import com.google.common.collect.Sets;
  *
  * @author David Graeff <david.graeff@web.de>
  */
+@Component(immediate = true, service = ThingHandlerFactory.class)
 public class MilightHandlerFactory extends BaseThingHandlerFactory {
     private static final Set<ThingTypeUID> SUPPORTED_THING_TYPES_UIDS = Sets
             .union(MilightBindingConstants.BRIDGE_THING_TYPES_UIDS, MilightBindingConstants.SUPPORTED_THING_TYPES_UIDS);
