+++ addons/binding/org.openhab.binding.milight/src/main/java/org/openhab/binding/milight/internal/discovery/MilightBridgeDiscovery.java
@@ -20,11 +20,13 @@ import java.util.concurrent.TimeUnit;
 import org.eclipse.smarthome.config.discovery.AbstractDiscoveryService;
 import org.eclipse.smarthome.config.discovery.DiscoveryResult;
 import org.eclipse.smarthome.config.discovery.DiscoveryResultBuilder;
+import org.eclipse.smarthome.config.discovery.DiscoveryService;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.openhab.binding.milight.MilightBindingConstants;
 import org.openhab.binding.milight.internal.MilightHandlerFactory;
 import org.openhab.binding.milight.internal.protocol.MilightDiscover;
 import org.openhab.binding.milight.internal.protocol.MilightDiscover.DiscoverResult;
+import org.osgi.service.component.annotations.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -34,6 +36,7 @@ import org.slf4j.LoggerFactory;
  *
  * @author David Graeff - Initial contribution
  */
+@Component(immediate = true, service = DiscoveryService.class)
 public class MilightBridgeDiscovery extends AbstractDiscoveryService implements DiscoverResult {
     private ScheduledFuture<?> backgroundFuture;
     private Logger logger = LoggerFactory.getLogger(MilightBridgeDiscovery.class);
