+++ addons/binding/org.openhab.binding.milightdev/ESH-INF/binding/binding.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<binding:binding id="milight"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
+        xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
+
+    <name>Milight Development Helper Binding</name>
+    <description>A binding for helping developing the Milight/Easybulb/compatible binding.</description>
+    <author>David Gräff</author>
+
+</binding:binding>
