+++ addons/binding/org.openhab.binding.milightdev/OSGI-INF/.gitignore
@@ -1,0 +1,1 @@
+*.xml
