+++ addons/binding/org.openhab.binding.milightdev/README.md
@@ -1,0 +1,4 @@
+# Milight/Easybulb/Limitless development helper binding
+
+Install this binding to help if a developer asked you to install it. It provides
+a simulated Milight bridge and will log received packets.
