+++ addons/binding/org.openhab.binding.milightdev/build.properties
@@ -1,0 +1,8 @@
+source.. = src/main/java/
+output.. = target/classes
+bin.includes = META-INF/,\
+               .,\
+               OSGI-INF/,\
+               ESH-INF/,\
+               about.html
+
