+++ addons/binding/org.openhab.binding.milightdev/src/main/java/org/openhab/binding/milight/internal/simulatedbridge/EmulatedV6Bridge.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.milight.internal.test;
+package org.openhab.binding.milight.internal.simulatedbridge;
 
 import java.io.IOException;
 import java.net.DatagramPacket;
