+++ addons/binding/pom.xml
@@ -53,6 +53,7 @@
     <module>org.openhab.binding.miele</module>
     <module>org.openhab.binding.mihome</module>
     <module>org.openhab.binding.milight</module>
+    <module>org.openhab.binding.milightdev</module>
     <module>org.openhab.binding.minecraft</module>
     <module>org.openhab.binding.netatmo</module>
     <module>org.openhab.binding.network</module>
